fix (files):fix some mistakes in the code
