+++ .local.env
@@ -1,4 +1,4 @@
-ACCESS_TOKEN="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiNjdjMDYzOWQ4MzZlZThiZTcwNjE5MGE5Iiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzYwMDI4NTYyfQ.rauH9_RtM1H9t6qbwRir88i2vr4UvTgvzD5mSF1Sz0Y"
+ACCESS_TOKEN="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiNjdjMDYzOWQ4MzZlZThiZTcwNjE5MGE5Iiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzYwMTk5MzQ5fQ.E61wk2cgsIEC0IihKE8vVS999r0OWh88WIE3e94F_6o"
 BASE_URL="https://flower.elevateegy.com"
 ALL_NOTIFICATION_URL="/api/v1/notifications/user"
 CLEAR_ALL_NOTIFICATION_URL="/api/v1/notifications/clear-all"
