+++ src/components/layout/header/index.tsx
@@ -2,7 +2,7 @@ import { cn } from "@lib/utils/cn.utils";
 import React from "react";
 import Notification from "./components/Notification";
 
-const Header = () => {
+export default function Header() {
     return (
         <header className={cn("flex h-10 w-full items-center justify-between bg-white p-4 text-red-500 shadow-md dark:text-white")}>
             {/* Title */}
@@ -12,6 +12,4 @@ const Header = () => {
             <Notification />
         </header>
     );
-};
-
-export default Header;
+}
