+++ src/lib/hooks/use-delete-all-notifications.hook.ts
@@ -1,6 +1,6 @@
 "use client";
 import { useMutation } from "@tanstack/react-query";
-import { DeleteAllNotifications } from "@lib/actions/delete-all-notifications.action";
+import { DeleteAllNotifications } from "@lib/actions/notifications/notifications.action";
 
 export function useDeleteAllNotifications() {
     const { mutateAsync, error, isPending, isSuccess } = useMutation({
