+++ src/lib/hooks/use-delete-single-notification.ts
@@ -1,7 +1,7 @@
 "use client";
 import { useMutation } from "@tanstack/react-query";
 
-import { DeleteNotification } from "@lib/actions/delete-single-notification.action";
+import { DeleteNotification } from "@lib/actions/notifications/notifications.action";
 
 export function useDeleteSingleNotification() {
     const { mutateAsync, error, isPending, isSuccess } = useMutation({
