+++ src/lib/hooks/use-mark-all-notifications-as-read.hook.ts
@@ -1,6 +1,6 @@
 "use client";
 import { useMutation } from "@tanstack/react-query";
-import { MarkAllNotificationsAsRead } from "@lib/actions/mark-all-notifications-as-read.action";
+import { MarkAllNotificationsAsRead } from "@lib/actions/notifications/notifications.action";
 
 export function useMarkAllNotificationsAsRead() {
     const { mutateAsync, error, isPending, isSuccess } = useMutation({
