+++ src/lib/types/mark-all-notifications-as-read.d.ts
@@ -1,4 +1,4 @@
-type MarkAllNotificationsAsReadRequest = {
+declare type MarkAllNotificationsAsReadRequest = {
     modifiedCount: number;
     unreadCount: number;
 };
